ResumeParser_pjt/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
black==23.11.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.104.1
flake8==6.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.48.1
nltk==3.8.1
numpy==1.25.2
packaging==23.2
pandas==2.1.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
plotly==5.17.0
pluggy==1.6.0
preshed==3.0.10
protobuf==4.25.8
pyarrow==21.0.0
pycodestyle==2.11.1
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.9.1
pyflakes==3.1.0
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.27.0
streamlit==1.28.1
tenacity==8.5.0
thinc==8.2.5
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
validators==0.35.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
weasel==0.3.4
websockets==15.0.1
zipp==3.23.0
